% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bulk.R
\name{Create_bulk}
\alias{Create_bulk}
\title{Creates html pages with maps and basic stats for multiple cities}
\usage{
Create_bulk(cities, country)
}
\arguments{
\item{cities}{A character vector of cities located in the same country.}

\item{country}{The name of a country.}
}
\value{
Nothing, used for its side effects (creates html files with maps and basic text)
}
\description{
Creates html pages with maps and basic stats for multiple cities
}
\examples{

Create_bulk(cities = c("Sibiu", "Alba Iulia"), country = "Romania")

}
