% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm_data.R
\name{Download_OSM}
\alias{Download_OSM}
\title{Download OSM data for whole countries.}
\usage{
Download_OSM(countries)
}
\arguments{
\item{countries}{The query to be searched.}
}
\value{
Used only for its side effects (downloads osm data).
}
\description{
Download OSM data for whole countries.
}
\examples{

Download_OSM(search = "Romania")

}
