% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gender.R
\name{FindGender}
\alias{FindGender}
\title{Find if a string refers to a person, and if so, finds gender}
\usage{
FindGender(search, language = "en", cache = TRUE, wait = 1)
}
\arguments{
\item{search}{The query to be searched.}

\item{language}{Two letters language codes to be passed to WikiData}
}
\value{
A data.frame (a tibble) with one row and four columns: Query, Gender, Description, and WikidataID
}
\description{
Find if a string refers to a person, and if so, finds gender
}
\examples{

FindGender(search = "Garibaldi", language = "it")

}
