% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_street_names.R
\name{Fix_street_names}
\alias{Fix_street_names}
\title{Starts a Shiny app that facilitates fixing categorisation}
\usage{
Fix_street_names(city, country)
}
\arguments{
\item{city}{The name of a city/municipality.}

\item{country}{The name of the country. Requested to ensure correct identification of city.}
}
\description{
Starts a Shiny app that facilitates fixing categorisation
}
\examples{
\dontrun{
Fix_street_names()
}
}
