% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm_data.R
\name{Get_boundary_by_id}
\alias{Get_boundary_by_id}
\title{Get boundary by id}
\usage{
Get_boundary_by_id(id, type = "way", cache = TRUE)
}
\arguments{
\item{id}{A numeric vector of length 1, must correspond to the id of a boundary object on OpenStreetMap.}

\item{type}{Defaults to "way".}

\item{cache}{Logical, defaults to TRUE. If TRUE, stores data in local subfolder data/cities/country_name/city_name.rds}
}
\value{
An sf polygon.
}
\description{
Get boundary by id
}
\examples{

Get_boundary_by_id(id = c(Arad = 45422208))

}
