% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm_data.R
\name{get_city_boundaries}
\alias{get_city_boundaries}
\title{Get city boundaries}
\usage{
get_city_boundaries(city, country, admin_level = 6,
  administrative = NULL, cache = TRUE)
}
\arguments{
\item{city}{The name of a city/municipality.}

\item{country}{The name of the country. Requested to ensure correct identification of city.}

\item{admin_level}{Defaults to 6. For more information see: https://wiki.openstreetmap.org/wiki/Tag:boundary\%3Dadministrative#10_admin_level_values_for_specific_countries}

\item{administrative}{Defaults to NULL. If TRUE, filters only boundaries recorded as administrative. For more information, see: https://wiki.openstreetmap.org/wiki/Tag\%3aboundary=administrative}

\item{cache}{Logical, defaults to TRUE. If TRUE, stores data in local subfolder data/cities/country_name/city_name.rds}
}
\value{
An sf polygon.
}
\description{
Get city boundaries
}
\examples{

get_city_boundaries(search = "Sibiu, Romania")

}
