% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_select_dataset.R
\name{mod_select_dataset_ui}
\alias{mod_select_dataset_ui}
\alias{mod_select_dataset}
\title{mod_select_dataset_ui and mod_select_dataset_server}
\usage{
mod_select_dataset_ui(id)

mod_select_dataset(input, output, session, dataset_folder_name)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{dataset_folder_name}{Datasets will be searched from the given folder.}
}
\description{
A shiny Module.
}
\keyword{internal}
