% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_street_names.R
\name{remove_first_word}
\alias{remove_first_word}
\title{Find if a string refers to a person, and if so, finds gender by searching Wikidata}
\usage{
remove_first_word(string)
}
\arguments{
\item{string}{A character string.}
}
\value{
A character string.
}
\description{
Find if a string refers to a person, and if so, finds gender by searching Wikidata
}
\examples{

remove_first_word(string = "via Garibaldi")

}
