% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bulk.R
\name{Create_bulk}
\alias{Create_bulk}
\title{Get city boundaries}
\usage{
Create_bulk(cities, country)
}
\arguments{
\item{cities}{A character vector of cities located in the same country.}

\item{country}{The name of a country.}
}
\value{
Nothing, used for its side effects (creates html files with maps and basic text)
}
\description{
Get city boundaries
}
\examples{

Create_bulk(cities = c("Sibiu", "Timișoara"), country = "Romania")

}
