% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm_data.R
\name{create_city_boundary_id_combo}
\alias{create_city_boundary_id_combo}
\title{Create a city/id combination for proper caching}
\usage{
create_city_boundary_id_combo(id, type, city = NULL, country = NULL,
  cache = TRUE)
}
\arguments{
\item{id}{A numeric vector of length 1, must correspond to the id of a boundary object on OpenStreetMap.}

\item{type}{One of either "way", "relation", or "node".}

\item{city}{The name of a city/municipality.}

\item{country}{The name of the country. Requested to ensure correct identification of city.}

\item{cache}{Logical, defaults to TRUE. If TRUE, stores data in local subfolder data/cities/country_name/city_name.rds}
}
\value{
A list, typically to be fed into `get_boundary_by_id()``
}
\description{
Create a city/id combination for proper caching
}
\examples{

create_city_boundary_id_combo(id = 45422208, type = "way", city = "Arad", country = "Romania")

# https://www.openstreetmap.org/relation/46663 
create_city_boundary_id_combo(id = 46663, type = "relation", city = "Trento", country = "Italy", )

}
