% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm_data.R
\name{download_OSM}
\alias{download_OSM}
\title{Download OSM data for whole countries.}
\usage{
download_OSM(countries, continent = "Europe")

download_OSM(countries, continent = "Europe")
}
\arguments{
\item{countries}{One or more country names. For details on the country names see http://download.geofabrik.de/}

\item{continent}{Defaults to `Europe`. Available options are: `africa`, `asia`, `australia-oceania`, `central-america`, `europe`, `north-america`, `south-america`. Only one continent at a time can be chosen.}

\item{countries}{One or more country names. For details on the country names see http://download.geofabrik.de/}

\item{continent}{Defaults to `Europe`. Available options are: `africa`, `asia`, `australia-oceania`, `central-america`, `europe`, `north-america`, `south-america`. Only one continent at a time can be chosen.}
}
\value{
Used only for its side effects (downloads osm data).

Used only for its side effects (downloads osm data).
}
\description{
Download OSM data for whole countries.

Download OSM data for whole countries.
}
\examples{

download_OSM(search = "Romania")


download_OSM(countries = "Romania")

}
