% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm_data.R
\name{Extract_roads}
\alias{Extract_roads}
\title{Extract shape files of roads from previously downloaded}
\usage{
Extract_roads(countries, export_rds = FALSE, export_csv = FALSE)
}
\arguments{
\item{countries}{The query to be searched.}

\item{export_rds}{Stores imported shape files as an rds file locally.}

\item{export_csv}{Stores imported shape files (excluding the geographic information) as a csv file locally.}
}
\value{
A data.frame with geographic data (sf).
}
\description{
Extract shape files of roads from previously downloaded
}
\examples{

Extract_roads(search = "Romania")

}
