% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm_data.R
\name{Get_city_boundaries}
\alias{Get_city_boundaries}
\title{Get city boundaries}
\usage{
Get_city_boundaries(city, country, cache = TRUE)
}
\arguments{
\item{city}{The name of a city/municipality.}

\item{country}{The name of the country. Requested to ensure correct identification of city.}

\item{cache}{Logical, defaults to TRUE. If TRUE, stores data in local subfolder data/cities/country_name/city_name.rds}
}
\value{
An sf polygon.
}
\description{
Get city boundaries
}
\examples{

Get_city_boundaries(search = "Sibiu, Romania")

}
