% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm_data.R
\name{Subset_roads}
\alias{Subset_roads}
\title{Keep only roads within a given boundary.}
\usage{
Subset_roads(boundary, roads)
}
\arguments{
\item{boundary}{An object typically created with `Get_city_boundaries()`}

\item{country}{The name of the country. Requested to ensure correct identification of city.}
}
\value{
A data frame of the sf class including all roads insidet the given boundary
}
\description{
Keep only roads within a given boundary.
}
\examples{

Subset_roads(city_boundary, roads)

}
